Move Trillium instrumented handler wrapper
