+++ aggregator_api/Cargo.toml
@@ -25,7 +25,6 @@ serde_test = "1.0.163"
 tracing = "0.1.37"
 trillium.workspace = true
 trillium-api.workspace = true
-trillium-macros = "0.0.4"
 trillium-opentelemetry.workspace = true
 trillium-router.workspace = true
 url = { version = "2.3.1", features = ["serde"] }
