+++ aggregator_core/Cargo.toml
@@ -57,12 +57,15 @@ tokio = { version = "1.28", features = ["full", "tracing"] }
 tokio-postgres = { version = "0.7.8", features = ["with-chrono-0_4", "with-serde_json-1", "with-uuid-1", "array-impls"] }
 tracing = "0.1.37"
 tracing-log = "0.1.3"
+trillium.workspace = true
+trillium-macros = "0.0.4"
+trillium-router.workspace = true
 url = { version = "2.3.1", features = ["serde"] }
 uuid = { version = "1.3.3", features = ["v4"] }
 
 [dev-dependencies]
 assert_matches = "1"
-async-std = { version = "1.12.0",  features = ["attributes"] }
+async-std = { version = "1.12.0", features = ["attributes"] }
 hyper = "0.14.26"
 janus_aggregator_core = { path = ".", features = ["test-util"] }
 janus_core = { workspace = true, features = ["test-util"] }
