+++ aggregator_core/src/lib.rs
@@ -5,6 +5,11 @@
 // https://github.com/rust-lang/rust-clippy/pull/9879
 #![allow(clippy::single_component_path_imports)]
 
+use tracing::{info, info_span, Instrument, Span};
+use trillium::{Conn, Handler, Status};
+use trillium_macros::Handler;
+use trillium_router::RouterConnExt;
+
 // We must import `rstest_reuse` at the top of the crate
 // https://docs.rs/rstest_reuse/0.5.0/rstest_reuse/#use-rstest_reuse-at-the-top-of-your-crate
 #[cfg(test)]
@@ -44,3 +49,34 @@ impl<P, const SEED_SIZE: usize> VdafHasAggregationParameter
 
 #[cfg(feature = "test-util")]
 impl VdafHasAggregationParameter for dummy_vdaf::Vdaf {}
+
+pub fn instrumented<H: Handler>(handler: H) -> impl Handler {
+    InstrumentedHandler(handler)
+}
+
+struct InstrumentedHandlerSpan(Span);
+
+#[derive(Handler)]
+struct InstrumentedHandler<H>(#[handler(except = [run, before_send])] H);
+
+impl<H: Handler> InstrumentedHandler<H> {
+    async fn run(&self, mut conn: Conn) -> Conn {
+        let route = conn.route().expect("no route in conn").to_string();
+        let method = conn.method();
+        let span = info_span!("endpoint", route, %method);
+        conn.set_state(InstrumentedHandlerSpan(span.clone()));
+        self.0.run(conn).instrument(span).await
+    }
+
+    async fn before_send(&self, conn: Conn) -> Conn {
+        if let Some(span) = conn.state::<InstrumentedHandlerSpan>() {
+            let _entered = span.0.enter();
+            let status = conn
+                .status()
+                .as_ref()
+                .map_or("unknown", Status::canonical_reason);
+            info!(status, "Finished handling request");
+        }
+        conn
+    }
+}
